backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
google-generativeai
pdf2image
pillow
